Indoor navigation
In this project titled “An Energy-efficient Indoor Navigation System using Smartphones” first various kind of indoor navigation methods were studied and then the need of a hybrid navigation system was found out.

To include the step counting technique in the hybrid indoor navigation system, a pedometer android app was using the raw values accelerometer sensor of the smartphone.

Lastly a survey was done at IIT BHU CSE department to identify different type of places and a rough map was drawn of the floor .Then the app was tested on that floor and accurate and good result were found.